move svelte files out of fastapi folders
