+++ app/package.json
@@ -26,6 +26,8 @@
   },
   "type": "module",
   "dependencies": {
-    "@observablehq/plot": "^0.6.10"
+    "@observablehq/plot": "^0.6.10",
+    "@types/d3": "^7.4.1",
+    "d3": "^7.8.5"
   }
 }
