+++ app/src/routes/+layout.svelte
@@ -1,8 +1,8 @@
 <script lang="ts">
   import "../app.postcss";
   import Icon from '@iconify/svelte';
-  import {displaySidebar} from "$lib/stores"
-  import NavigatonSidebar from '../components/NavigatonSidebar.svelte';
+  import {sidebarViewStatus} from "$lib"
+  import NavigatonSidebar from '../lib/components/NavigatonSidebar.svelte';
   import { goto } from "$app/navigation";
   $: query = ""
 
@@ -12,13 +12,13 @@
 </script>
 
 
-{#if $displaySidebar}
+{#if $sidebarViewStatus}
   <NavigatonSidebar/>
 {/if}
 
 <div class="w-screen bg-indigo-700 py-2 text-white inline-flex drop-shadow-lg px-4">
 
-  <button on:click={() => $displaySidebar = !$displaySidebar} class="pr-2">
+  <button on:click={() => sidebarViewStatus.toggleSidebar()} class="pr-2">
     <Icon icon="material-symbols:density-medium-rounded" height=32 width=32/>
   </button>
