+++ app/src/routes/search/[query]/+page.server.ts
