+++ app/src/routes/search/[query]/+page.svelte
