+++ app/svelte.config.js
