+++ preordain/config.py
@@ -52,7 +52,7 @@ FOLDER_PATH = "./preordain/v1/images/{type}/{set}/"
 
 # V2
 IMG_FOLDER_PATH_V2 = "./preordain/v2/images/files/{scryfall_uri}"
-V2_DASHBOARD = "preordain/v2/static/preordain/dist"
+V2_DASHBOARD = "./app/dist/"
 
 if TESTING:
     DB_NAME = Secret("test_" + str(DB_NAME))
@@ -79,18 +79,16 @@ if DASHBOARD:
     # TODO: This should build the static files, which will then serve right after.
     if not os.path.exists(V2_DASHBOARD):
         log.warning("Dashboard failed to load! Building right now...")
-        subprocess.Popen(
-            "npm run dev --prefix preordain/v2/static/preordain/", shell=True
-        )
-
-    # API_CONFIG[
-    #     "description"
-    # ] += "Dashboard Enabled, the dashboard should be available on the /dash part."
-    # API_CONFIG["routes"].append(
-    #     Mount(
-    #         "/v2/dash",
-    #         app=StaticFiles(directory="preordain/v2/static/preordain/dist/", html=True),
-    #         name="dashboard",
-    #     )
-    # )
-    # subprocess.Popen(["npm", "run", "dev", "--prefix", "preordain/v2/static/preordain/"], shell=True)
+        subprocess.Popen("npm run dev --prefix ./app", shell=True)
+# --prefix preordain/v2/static/preordain/
+# API_CONFIG[
+#     "description"
+# ] += "Dashboard Enabled, the dashboard should be available on the /dash part."
+# API_CONFIG["routes"].append(
+#     Mount(
+#         "/v2/dash",
+#         app=StaticFiles(directory="preordain/v2/static/preordain/dist/", html=True),
+#         name="dashboard",
+#     )
+# )
+# subprocess.Popen(["npm", "run", "dev", "--prefix", "preordain/v2/static/preordain/"], shell=True)
