--- preordain/v2/static/preordain/src/lib/stores.ts
@@ -1,3 +1,0 @@
-import { writable } from "svelte/store";
-
-export const displaySidebar = writable(false)
